pyproject.toml
[tool.poetry]
name = "sf-dune-utils"
version = "0.1.0"
description = "Sync SuperToken holder data to Dune Analytics"
authors = ["Superfluid"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.0"
dune-client = "^1.3.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
